+++ packages/definition-compiler/src/render/components/Whitespace.ts
@@ -1,0 +1,11 @@
+import { defineComponent, compile } from 'vue';
+
+export default defineComponent({
+    props: {
+        t: {
+            type: String,
+            required: true,
+        },
+    },
+    render: compile(`{{ t }}`),
+});
